fix: RequestParam 및 Slice Service 코드 개선

+++ src/main/java/com/sejong/sejongpeer/domain/study/repository/StudyRepositoryImpl.java
@@ -24,7 +24,7 @@ public class StudyRepositoryImpl implements StudyRepositoryCustom {
                         .selectFrom(study)
                         .where(ltStudyId(lastId))
                         .orderBy(study.id.desc())
-                        .limit(size + 1)
+                        .limit((long) size + 1)
                         .fetch();
         return checkLastPage(size, studies);
     }
